+++ src/main.ts
@@ -2,6 +2,7 @@ import {createHighlighterCore} from 'shiki/core';
 import javascript from 'shiki/langs/javascript.mjs';
 import css from 'shiki/langs/css.mjs';
 import htm from 'shiki/langs/html.mjs';
+import json from 'shiki/langs/json.mjs';
 import loadWasm from 'shiki/wasm';
 import monokai from 'shiki/themes/monokai.mjs';
 import nord from 'shiki/themes/nord.mjs';
@@ -61,6 +62,7 @@ const registerWiki = async (monaco: typeof Monaco, parserConfig: Config | boolea
 			javascript,
 			css,
 			htm,
+			json,
 		],
 		themes: [
 			monokai,
@@ -68,10 +70,11 @@ const registerWiki = async (monaco: typeof Monaco, parserConfig: Config | boolea
 		],
 		loadWasm,
 	});
-	monaco.languages.register({id: 'wikitext', aliases: ['wiki', 'mediawiki']});
+	monaco.languages.register({id: 'wikitext', aliases: ['Wikitext', 'mediawiki', 'MediaWiki', 'wiki']});
 	monaco.languages.register({id: 'javascript', aliases: ['JavaScript', 'js']});
 	monaco.languages.register({id: 'css', aliases: ['CSS']});
 	monaco.languages.register({id: 'html', aliases: ['HTML', 'htm', 'xhtml']});
+	monaco.languages.register({id: 'json', aliases: ['JSON']});
 	shikiToMonaco(highlighter, monaco);
 	monaco.languages.setLanguageConfiguration('wikitext', config);
 	monaco.languages.registerCompletionItemProvider('wikitext', completion(monaco));